Refactor orchestrators around shared base class
